Update Code of Conduct 💣
This refinement of the CoC is based on my experiences from other
projects. A series of edits might follow.
As it has a low contributor count, I use this project as a playground to
develop a good CoC which does not focus on identity politics in order to
be a useful tool even with a hate mob pushing in.

My issues with the contributor covenant:

- I don't like its authoritarian tone, I want something more welcoming,
  improvement oriented. That'll be addressed next.

- Listing minorities explicitly is en vogue, but does not help to define
  acceptable behavior. Looking at prominent projects, this property has
  actually been gamed by harassers to say "but check my privilege, i am
  $x myself!", which in turn ended up with a non-privilege competition,
  generating more toxic waste.
  Why don't we just say *there is no acceptable reason for harassment*?

- The list of examples again tries to be a deny-list, where it is
  actually just a list of... examples. Let's simplify that list, as it
  mostly contains duplicates anyway.
  Every explicitly mentioned action provides additional attack surface
  for straw-man discussions.